refactor encrypt rock auth in list api
